backend/src/requirements.txt
boto3
requests
python-dateutil
pytz
PyYAML
cryptography
